fix: make search and support available in zen mode
